Refactor html to pdf conversion and add error handling.
Previously running the wkhtmltopdf command could fail, either by hanging and timing out or by returning a nonzero return code. Both cases were silently logged but did not produce an error. This could result in some strange cases where faxing was attempting to fax out an empty file.

+++ src/main/java/oscar/eform/actions/FaxAction.java
@@ -8,11 +8,7 @@
  */
 package oscar.eform.actions;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.ArrayList;
+import java.io.*;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.HashSet;
@@ -25,6 +21,7 @@ import org.apache.log4j.Logger;
 import org.oscarehr.common.dao.EFormDataDao;
 import org.oscarehr.common.dao.FaxConfigDao;
 import org.oscarehr.common.dao.FaxJobDao;
+import org.oscarehr.common.exception.HtmlToPdfConversionException;
 import org.oscarehr.common.model.EFormData;
 import org.oscarehr.common.model.FaxConfig;
 import org.oscarehr.common.model.FaxJob;
@@ -33,8 +30,6 @@ import org.oscarehr.util.SpringUtils;
 import org.oscarehr.util.WKHtmlToPdfUtils;
 
 import oscar.OscarProperties;
-
-import com.lowagie.text.DocumentException;
 
 import com.itextpdf.text.pdf.PdfReader;
 
@@ -87,61 +82,56 @@ public final class FaxAction {
 	}
 
 	/**
-	 * This method will take eforms and send them to a PHR.
-	 * @throws DocumentException 
+	 * Prepares eForm fax files and places them in the outgoing faxes location.
+	 *
+	 * @param numbers     the fax numbers to send to
+	 * @param formId      the fdid of the form to send
+	 * @param providerId  the provider number to record in the faxJob
+	 *
+	 * @throws IOException
+	 * @throws HtmlToPdfConversionException
 	 */
-	public void faxForms(String[] numbers, String formId, String providerId) throws DocumentException {
-		
+	public void faxForms(String[] numbers, String formId, String providerId) throws IOException, HtmlToPdfConversionException
+	{
+		String faxFileLocation = OscarProperties.getInstance().getProperty(
+				"fax_file_location", System.getProperty("java.io.tmpdir"));
+		String temporaryFileLocation = OscarProperties.getInstance().getProperty("DOCUMENT_DIR") + "/";
 		File tempFile = null;
 
 		try {
-			logger.info("Generating PDF for eform with fdid = " + formId);
+			logger.info("Generating PDF for eForm with fdid = " + formId);
 
 			String pdfFile = "EForm." + formId + System.currentTimeMillis();
-			tempFile = File.createTempFile(pdfFile, ".pdf");
-			//tempFile.deleteOnExit();
+			tempFile = File.createTempFile(pdfFile, ".pdf", new File(temporaryFileLocation));
 
 			// convert to PDF
 			String viewUri = localUri + formId;
+			logger.info("Converting eForm content to pdf. Target file: " + tempFile.getCanonicalPath());
 			WKHtmlToPdfUtils.convertToPdf(viewUri, tempFile);
-			logger.info("Writing pdf to : "+tempFile.getCanonicalPath());
-			
-			// Removing all non digit characters from fax numbers.
-			for (int i = 0; i < numbers.length; i++) { 
-				numbers[i] = numbers[i].trim().replaceAll("\\D", "");
-			}
-			ArrayList<String> recipients = new ArrayList<String>(Arrays.asList(numbers));
-			
-			String path = OscarProperties.getInstance().getProperty("DOCUMENT_DIR") + "/";
-			FileUtils.copyFile(tempFile, new File(path + tempFile.getName()));
-			
-			// Removing duplicate phone numbers.
-			recipients = new ArrayList<String>(new HashSet<String>(recipients));
-			FileOutputStream fos;
-			for (int i = 0; i < recipients.size(); i++) {					
-				
-			    String faxNo = recipients.get(i).trim().replaceAll("\\D", "");
-			    if (faxNo.length() < 7) { 
-					throw new DocumentException("Document target fax number '"+faxNo+"' is invalid."); 
-				}
 
-				//String demo = req.getParameter("demographic_no");
-				FaxJobDao faxJobDao = SpringUtils.getBean(FaxJobDao.class);
-				FaxConfigDao faxConfigDao = SpringUtils.getBean(FaxConfigDao.class);
-				List<FaxConfig> faxConfigs = faxConfigDao.findAll(null, null);
-				//String provider_no = 
-				//	LoggedInInfo.getLoggedInInfoFromSession(req).getLoggedInProviderNo();
+			// Removing duplicate phone numbers.
+			HashSet<String> recipients = new HashSet<>(Arrays.asList(numbers));
+
+			FileOutputStream fos;
+			FaxJobDao faxJobDao = SpringUtils.getBean(FaxJobDao.class);
+			FaxConfigDao faxConfigDao = SpringUtils.getBean(FaxConfigDao.class);
+			List<FaxConfig> faxConfigs = faxConfigDao.findAll(null, null);
+			for (String recipient : recipients) {
+				String fileNameFormat = "EForm-" + formId + "." + System.currentTimeMillis();
+				String pdfFileName = String.format("%s%s%s.pdf", faxFileLocation, File.separator, fileNameFormat);
+				String txtFileName = String.format("%s%s%s.txt", faxFileLocation, File.separator, fileNameFormat);
+
+				// Copying the fax pdf.
+				FileUtils.copyFile(tempFile, new File(pdfFileName));
 
 				FaxJob faxJob;
-				boolean validFaxNumber = false;
-			                
+
 				for( FaxConfig faxConfig : faxConfigs ) {
 
-					PdfReader pdfReader = new PdfReader(path + tempFile.getName());
+					PdfReader pdfReader = new PdfReader(temporaryFileLocation + tempFile.getName());
 
 					faxJob = new FaxJob();
-					faxJob.setDestination(faxNo);
-					//faxJob.setFax_line(faxNumber);
+					faxJob.setDestination(recipient);
 					faxJob.setFax_line(null);
 					faxJob.setFile_name(tempFile.getName());
 					faxJob.setUser(faxConfig.getFaxUser());
@@ -149,48 +139,39 @@ public final class FaxAction {
 					faxJob.setStamp(new Date());
 					faxJob.setStatus(FaxJob.STATUS.SENT);
 					faxJob.setOscarUser(providerId);
-					//faxJob.setDemographicNo(Integer.parseInt(demo));
 					faxJob.setDemographicNo(null);
 
 					faxJobDao.persist(faxJob);
-					validFaxNumber = true;
 					break;
 				}
-
-			    String tempName = "EForm-" + formId + "." + System.currentTimeMillis();
-				String tempPath = OscarProperties.getInstance().getProperty(
-						"fax_file_location", System.getProperty("java.io.tmpdir"));
-				
-				String tempPdf = String.format("%s%s%s.pdf", tempPath, File.separator, tempName);
-				String tempTxt = String.format("%s%s%s.txt", tempPath, File.separator, tempName);
-				
-				// Copying the fax pdf.
-				FileUtils.copyFile(tempFile, new File(tempPdf));
 				
 				// Creating text file with the specialists fax number.
-				fos = new FileOutputStream(tempTxt);				
+				fos = new FileOutputStream(txtFileName);
 				PrintWriter pw = new PrintWriter(fos);
-				pw.println(faxNo);
+				pw.println(recipient);
 				pw.close();
 				fos.close();
 				
-				// A little sanity check to ensure both files exist.
-				if (!new File(tempPdf).exists() || !new File(tempTxt).exists()) {
-					throw new DocumentException("Unable to create files for fax of eform " + formId + ".");
-				}		
-				if (skipSave) {
-		        	 EFormDataDao eFormDataDao=(EFormDataDao) SpringUtils.getBean("EFormDataDao");
-		        	 EFormData eFormData=eFormDataDao.find(Integer.parseInt(formId));
-		        	 eFormData.setCurrent(false);
-		        	 eFormDataDao.merge(eFormData);
+				// A little sanity check to ensure the file exists.
+				if (!new File(txtFileName).exists()) {
+					throw new IOException("Unable to create fax file for eForm " + formId + ".");
 				}
 			}
-			// Removing the consulation pdf.
-			tempFile.delete();			
-						
-		} catch (IOException e) {
-			MiscUtils.getLogger().error("Error converting and sending eform. id="+formId, e);
-		} 
+
+			if (skipSave) {
+				EFormDataDao eFormDataDao=(EFormDataDao) SpringUtils.getBean("EFormDataDao");
+				EFormData eFormData=eFormDataDao.find(Integer.parseInt(formId));
+				eFormData.setCurrent(false);
+				eFormDataDao.merge(eFormData);
+			}
+		} finally
+		{
+			// Removing the temp pdf.
+			if (tempFile != null)
+			{
+				tempFile.delete();
+			}
+		}
 	}
 
 }
